feat(File) : File 도메인 생성 rename : File 관련 controller, service, entity 이동 / S3UploadService -> FileServiceImpl 이름 변경 feat(FileService) : FileService interface 생성 feat(UploadFileResponseDto) : 파일 업로드 ResponseDto 생성 feat(SecurityConfig) : 파일 업로드 설정 추가 refactor(FileServiceImpl) : 반환 타입 변경 String -> UploadFileResponseDto refactor(FileController) : fileUpload에 PostMapping MediaType 추가
 #47
