+++ src/main/java/com/creavispace/project/domain/file/dto/response/UploadFileResponseDto.java
@@ -1,0 +1,14 @@
+package com.creavispace.project.domain.file.dto.response;
+
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+
+@Getter
+@Builder
+@NoArgsConstructor
+@AllArgsConstructor
+public class UploadFileResponseDto {
+    private String url;
+}
